Sending Gameplay Events
Gameplay Event can be send and received. There are two nodes:
Wait Gameplay Event: which will wait for Gameplay Event to be send with a matching Event Tag. It can only be used in Gameplay Ability class.
Send Gameplay Event to Actor: This will send a Gameplay Event with a Event Tag and any Wait Gameplay Event node will receive it if it has a matching Event Tag.